Fix postgres types

+++ src/utils/wrappa/postgres/query.ts
@@ -134,8 +134,8 @@ const queryAggregatedDailyTimestampRange = async (
     date_trunc('day', ts) AS ts, 
     CAST(SUM(total_deposited_usd) AS BIGINT) AS total_deposited_usd, 
     CAST(SUM(total_withdrawn_usd) AS BIGINT) AS total_withdrawn_usd, 
-    CAST(SUM(total_deposit_txs) AS BIGINT) AS total_deposit_txs, 
-    CAST(SUM(total_withdrawal_txs) AS BIGINT) AS total_withdrawal_txs 
+    CAST(SUM(total_deposit_txs) AS INT) AS total_deposit_txs, 
+    CAST(SUM(total_withdrawal_txs) AS INT) AS total_withdrawal_txs 
   FROM 
     bridges.hourly_aggregated
   WHERE
